Factor our reusable tasks from FidesmoPlugin
